Fix remote linters errors
