+++ index.html
@@ -103,7 +103,7 @@
 <footer class="fixed-bottom border bg-light-transparent m-0 px-3 pt-2">
     <p class="text-20 mb-0"> &copy;2022 All copyrights are reserved </p>
 </footer>
-<script type="module" src="./index.js" ></script>
+<script src="./index.js" type="module"></script>
 </body>
 </html>
