--- modules/dom_body_elements.js
@@ -1,19 +1,0 @@
-const bookForm = document.getElementById('book-form');
-const bookList = document.querySelector('.books');
-const msgBlock = document.getElementById('msg-block');
-const addMsgBlock = document.getElementById('add-msg-block');
-const navList = document.querySelector('#nav-list');
-const navAddNew = document.querySelector('#nav-add-new');
-const navContact = document.querySelector('#nav-contact');
-const bookListSection = document.querySelector('.books-list');
-const addNewSection = document.querySelector('.add-new-book');
-const contactSection = document.querySelector('.add-contact-info');
-const hamburgerMenu = document.querySelector('.hamburger');
-const navBar = document.querySelector('nav');
-const navLinks = document.querySelectorAll('#nav-links li');
-export {
-  bookForm, bookList, msgBlock,
-  addMsgBlock, navList, navAddNew,
-  navContact, bookListSection, addNewSection,
-  contactSection, hamburgerMenu, navBar, navLinks,
-};
